* Continue refactoring menu management by introducing a PoseViewListener mechanism. That i will most certainly use for the other controls but only the File menu uses it for now.
* Consequently fixed updating bugs of the File menu that wouldn't change
depending on the context (root, trash, dir in trash and printers dirs)
and selection changes.

The new File menu is now fully equipped to work in the root folder thus
allowing to get rid of the VolumeWindow BContainerWindow variant (just 
need to finish the context menu part of VolumeWindow).
Depending on where you started navigating in single window mode, Tracker 
was using either VolumeWindow or the generic BContainerWindow (which
wasn't behaving correctly).

Temporarily disabled the 'Open With...' submenu as there's a problem 
that i need to take care of and i want to commit the current working 
code.

* Got rid of 
fSelectionChangedHook boolean and associated mess, it's useless as 
there's already an empty impl method in the base class.

* Set explicit min width of the title view to 0.0 so that it plays well
with layouting and the window's B_AUTO_UPDATE_SIZE_LIMITS.


git-svn-id: file:///srv/svn/repos/haiku/haiku/branches/developer/aldeck/tracker_layout@39086 a95241bf-73f2-0310-859d-f6bbb57e9c96